iio: adc: ad4630: remove ad463x compatible
Having a catch all compatible to detect the device is uncommon and not
the typical way of handling things where an explicit compatible is
required. As this is no longer needed, let's remove it.

Note that we still do a sanity grade check and print a warning in case
of mismatch.

Signed-off-by: Nuno Sa <nuno.sa@analog.com>